requirements.txt
python-telegram-bot[job-queue]==20.3
python-dotenv==1.0.0
requests==2.31.0
matplotlib==3.7.1
numpy==1.24.3